% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_num_mat.R
\name{check_num_mat}
\alias{check_num_mat}
\title{Checks that all elements in a matrix are numeric}
\usage{
check_num_mat(mat2check)
}
\value{
Returns TRUE for yes, FALSE for no
}
\description{
Checks that all elements in a matrix are numeric
}
