% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{cross_validation}
\alias{cross_validation}
\title{cross_validation}
\usage{
cross_validation(
  training_dataset = training_dataset,
  seed = seed,
  groups = groups,
  nfolds = nfolds
)
}
\arguments{
\item{training_dataset}{the training data}

\item{seed}{initialize a random number generator with a specific seed value}

\item{groups}{information on how the stratified data will be grouped (name of column, index of column, vector)}

\item{nfolds}{number of folds in the data if using split_data for cross validation}
}
\description{
this function is used to create folds of the training data for training the model
}
\examples{
number_of_participants <- 10
covariance_matrix <- diag(2)
outcome_column <- 1
means <- c(0,0)
seed <- set.seed(7)
groups <- rep(1:2, each = 5) #information on how the stratified data will be grouped
nfolds <- 5
data <- simulate_data(number_of_participants,covariance_matrix,outcome_column,means)
randomized_data <- randomize(training_dataset=data,seed=seed)
stratified_data <- stratify_data(randomized_data=randomized_data, groups=groups)
folded_data <- split_data(stratified_data=stratified_data, nfolds=nfolds)
}
\keyword{cross_validation}
