% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parse.rf.dat}
\alias{parse.rf.dat}
\title{Parse input data and return user params, datatype, unformatted data}
\usage{
parse.rf.dat(data.path, dat.type = c("ds", "pm"), outcome.var = NULL)
}
\value{
Returns missing data info, unformatted data, user inputs, data type in a list
}
\description{
Parse input data and return user params, datatype, unformatted data
}
