% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_rf_dat.R
\name{parse_rf_dat}
\alias{parse_rf_dat}
\title{Parse input data and return user params, datatype, unformatted data}
\usage{
parse_rf_dat(data.path, dat.type = c("ds", "pm"), rm.dat = c("rm",
  "cs"), outcome.var = NULL)
}
\value{
Returns missing data info, unformatted data, user inputs, data type in a list
}
\description{
Parse input data and return user params, datatype, unformatted data
}
