% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize.R
\name{randomize}
\alias{randomize}
\title{randomize}
\usage{
randomize(data, seed = NULL)
}
\arguments{
\item{data}{dataframe of the full data read in by the read_file function}

\item{seed}{initialize a random number generator with a specific seed value}
}
\description{
this function provides a randomized dataset to be used by the stratify_data function.
}
\examples{
simulate_data(number_of_participants=10, covariance_matrix=diag(2), outcome_column=1, means=c(0,0))
randomized_data <- randomize(data=data, seed = 7)
}
\keyword{randomize}
