% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file.R
\name{read_file}
\alias{read_file}
\title{read_file}
\usage{
read_file(file_path = file_path)
}
\arguments{
\item{file_path}{path to the input dataframe}
}
\description{
this function reads in an input file to be used to create the input dataframe.
}
\examples{
file_path = "path/to/file"
data = read_file(file_path=file_path)
}
\keyword{read_file}
