% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RF_optimal_train.R
\name{RF_optimal_train}
\alias{RF_optimal_train}
\title{RF_optimal_train}
\usage{
RF_optimal_train(
  data = training_dataset,
  formula = y ~ x,
  Optimal_RF_Parameters = NULL,
  model_type = "rfsrc"
)
}
\arguments{
\item{formula}{formula describing the model fit}

\item{Optimal_RF_Parameters}{returned from RF_test function}

\item{model_type}{which type of rfsrc to use}

\item{training_dataset}{the training data}
}
\description{
this function trains the optimal random forest parameters with training folds
}
\examples{
number_of_training_participants <- 10
covariance_matrix <- diag(2)
outcome_column <- 1
means <- c(0,0)
training_dataset <- simulate_data(number_of_training_participants,covariance_matrix,outcome_column,means)
optimal_random_forest <- RF_optimal_train(data=training_dataset,formula=y~x,Optimal_RF_Parameters=NULL,model_type="rfsrc")
}
\keyword{RF_optimal_train}
