% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RF_test.R
\name{RF_test}
\alias{RF_test}
\title{RF_test}
\usage{
RF_test(
  object = random_forest,
  newdata = testing_dataset,
  proximity_flag = TRUE
)
}
\arguments{
\item{proximity_flag}{boolean; whether or not proximity should be calculated}

\item{random_forest}{the resulting object from RF_train function}

\item{testing_dataset}{the resulting object from simulate_data function}
}
\description{
this function trains a random forest model with training data folds
}
\examples{
number_of_training_participants <- 10
number_of_testing_participants <- 10
covariance_matrix <- diag(2)
outcome_column <- 1
means <- c(0,0)
training_dataset <- simulate_data(number_of_training_participants,covariance_matrix,outcome_column,means)
mtry <- 1
nodesize <- 1
formula <- y~x
model_type <- rfsrc()
random_forest <- RF_Train(training_dataset,formula,mtry,nodesize,model_type)
testing_dataset <- simulate_data(number_of_testing_participants,covariance_matrix,outcome_column,means)
Optimal_RF_Parameters <- RF_Test(object=random_forest,newdata=testing_dataset,proximity=TRUE)
}
\keyword{RF_test}
