% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RF_train.R
\name{RF_train}
\alias{RF_train}
\title{RF_train}
\usage{
RF_train(
  data = training_dataset,
  formula = y ~ x,
  mtry = NULL,
  nodesize = NULL,
  model_type = "rfsrc"
)
}
\arguments{
\item{data}{the training dataset}

\item{ntrain_folds}{number of training folds}
}
\description{
this function trains a random forest model with training data folds
}
\examples{
number_of_training_participants <- 10
covariance_matrix <- diag(2)
outcome_column <- 1
means <- c(0,0)
training_dataset <- simulate_data(number_of_training_participants,covariance_matrix,outcome_column,means)
random_forest <- RF_Train(data=training_dataset,formula=y~x,mtry=NULL,nodesize=NULL,model_type="rfsrc")
}
\keyword{RF_train}
