% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{simulate data}
\usage{
simulate_data(
  number_of_participants = 3,
  covariance_matrix = diag(2),
  outcome_column = 1,
  means = c(0, 0)
)
}
\arguments{
\item{number_of_participants}{Provide the number of participants in the dataset. Default is 3.}

\item{covariance_matrix}{Matrix representing the relationship between the variables. Default is an identity matrix with 2 variables.}

\item{outcome_column}{Indicates what the outcome column is to be specified. Default is first column.}

\item{means}{Indicates the means of the matrix. Default is c(0,0).}
}
\description{
this function provides a simulated data set for the RF unit test.
}
\examples{
simulate_data(number_of_participants=3, covariance_matrix=diag(2), outcome_column=1, means=c(0,0))
}
\keyword{simulate}
