% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data.R
\name{split_data}
\alias{split_data}
\title{split_data}
\usage{
split_data(stratified_data, holdout_fraction, nfolds)
}
\arguments{
\item{stratified_data}{dataframe of the stratified data returned by the stratify_by function}

\item{holdout_fraction}{fraction of data that will be heldout for training the model, if using split_data for holdout}

\item{nfolds}{number of folds in the data if using split_data for cross validation}
}
\description{
this function provides a split dataset to be used by the cross_validation function.
}
\examples{
simulate_data(number_of_participants=10, covariance_matrix=diag(2), outcome_column=1, means=c(0,0))
randomized_data <- randomize(data=data, seed = 7)
stratified_data <- stratify_data(randomized_data=randomized_data,groups=groups)
train_and_test <- split_data(stratified_data=stratified_data,holdout_fraction=0.8)
folded_data <- split_data(stratified_data=train_and_test$training_dataset,nfolds=5)
}
\keyword{split}
