% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify_data.R
\name{stratify_data}
\alias{stratify_data}
\title{stratify_by}
\usage{
stratify_data(randomized_data, groups)
}
\arguments{
\item{randomized_data}{dataframe of the randomized data returned by the randomize function}

\item{groups}{information on how the stratified data will be grouped (name of column, index of column, vector)}
}
\description{
this function provides a stratified dataset to be used by the split_data function.
}
\examples{
simulate_data(number_of_participants=10, covariance_matrix=diag(2), outcome_column=1, means=c(0,0))
randomized_data <- randomize(data=data, seed = 7)
stratified_data <- stratify_data(randomized_data=randomized_data,groups=groups)
}
\keyword{stratify}
