% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subtype_ID.R
\name{Subtype_ID}
\alias{Subtype_ID}
\title{Subtype_ID}
\usage{
Subtype_ID(
  proximity_matrix = proximity_matrix,
  mode = mode,
  weighted = weighted,
  diag = diag
)
}
\arguments{
\item{proximity_matrix}{the resulting proximity matrix from RF_prediction}

\item{mode}{specifies how igraph should interpret the supplied matrix}

\item{weighted}{whether or not to create a weighted graph from supplied matrix}

\item{diag}{whether or not to include the diagonal of the supplied matrix in the calculation}
}
\description{
this function takes the proximity matrix from RF_optimal_test and creates a proximity graph, then utilizes infomap for community detection
}
\examples{
proximity_matrix <- RF_prediction$proximity
mode = "undirected"
weighted = TRUE
diag = FALSE
subtypes = Subtype_ID(proximity_matrix=proximity_matrix, mode=mode, weighted=weighted, diag=diag)
}
\keyword{Subtype_ID}
